Readonly properties feature (#144)
Added readonly properties.
Closes #139
