+++ bgw-core/src/main/kotlin/tools/aqua/bgw/components/container/LinearLayout.kt
@@ -22,8 +22,8 @@ package tools.aqua.bgw.components.container
 import tools.aqua.bgw.components.gamecomponentviews.GameComponentView
 import tools.aqua.bgw.components.uicomponents.Orientation
 import tools.aqua.bgw.core.*
-import tools.aqua.bgw.observable.DoubleProperty
-import tools.aqua.bgw.observable.Property
+import tools.aqua.bgw.observable.properties.DoubleProperty
+import tools.aqua.bgw.observable.properties.Property
 import tools.aqua.bgw.visual.Visual
 
 /**
