+++ bgw-core/src/main/kotlin/tools/aqua/bgw/components/uicomponents/LabeledUIComponent.kt
@@ -20,9 +20,9 @@
 package tools.aqua.bgw.components.uicomponents
 
 import tools.aqua.bgw.core.Alignment
-import tools.aqua.bgw.observable.BooleanProperty
-import tools.aqua.bgw.observable.Property
-import tools.aqua.bgw.observable.StringProperty
+import tools.aqua.bgw.observable.properties.BooleanProperty
+import tools.aqua.bgw.observable.properties.Property
+import tools.aqua.bgw.observable.properties.StringProperty
 import tools.aqua.bgw.util.Font
 import tools.aqua.bgw.visual.Visual
