+++ bgw-core/src/main/kotlin/tools/aqua/bgw/components/uicomponents/ListView.kt
@@ -23,7 +23,7 @@ import tools.aqua.bgw.core.DEFAULT_LISTVIEW_HEIGHT
 import tools.aqua.bgw.core.DEFAULT_LISTVIEW_WIDTH
 import tools.aqua.bgw.observable.ObservableArrayList
 import tools.aqua.bgw.observable.ObservableList
-import tools.aqua.bgw.observable.Property
+import tools.aqua.bgw.observable.properties.Property
 import tools.aqua.bgw.util.Font
 import tools.aqua.bgw.visual.Visual
