+++ bgw-core/src/main/kotlin/tools/aqua/bgw/components/uicomponents/ProgressBar.kt
@@ -21,8 +21,8 @@ package tools.aqua.bgw.components.uicomponents
 
 import tools.aqua.bgw.core.DEFAULT_PROGRESSBAR_HEIGHT
 import tools.aqua.bgw.core.DEFAULT_PROGRESSBAR_WIDTH
-import tools.aqua.bgw.observable.DoubleProperty
-import tools.aqua.bgw.observable.Property
+import tools.aqua.bgw.observable.properties.DoubleProperty
+import tools.aqua.bgw.observable.properties.Property
 import tools.aqua.bgw.util.Font
 import tools.aqua.bgw.visual.Visual
 import java.awt.Color
