+++ bgw-core/src/main/kotlin/tools/aqua/bgw/components/uicomponents/ToggleButton.kt
@@ -21,8 +21,8 @@ package tools.aqua.bgw.components.uicomponents
 
 import tools.aqua.bgw.core.DEFAULT_TOGGLE_BUTTON_HEIGHT
 import tools.aqua.bgw.core.DEFAULT_TOGGLE_BUTTON_WIDTH
-import tools.aqua.bgw.observable.BooleanProperty
-import tools.aqua.bgw.observable.Property
+import tools.aqua.bgw.observable.properties.BooleanProperty
+import tools.aqua.bgw.observable.properties.Property
 import tools.aqua.bgw.util.Font
 import tools.aqua.bgw.visual.Visual
